implementation_nn_from_scratch
In this project, each layer of FullyConnected, Activation(Relu, Tanh, Sigmoid), Softmax, Linear and Forward and Backward functions are written, and then using these layers, set up a network and train and test it. The network is implemented using Dataset Mnist. At the end of the Test section, the accuracy (99.88) of the grid is printed, and Matplotlib is also used to display the input Image in this section.